Adding user controller, adding JWT handling and authorization

+++ src/controller/authController.js
@@ -2,11 +2,12 @@ const md5 = require("md5");
 const UserModel = require("../model/user");
 const handleError = require("../utils/handleError");
 const { INGRESS_SETTINGS_OPTIONS } = require("firebase-functions/v1");
+const jwt = require("jsonwebtoken");
 
 
 class AuthController {
     static async register(req, res) {
-        const {name, password, address, place, country, street, long, lat, province} = req.body;
+        const {name, email, password, address, place, country, street, long, lat, province} = req.body;
 
         try {
             const duplicateUser = UserModel.findOne({email: email});
@@ -14,8 +15,10 @@ class AuthController {
                 throw new Error("User already exists.");
             }
 
+            const token = jwt.sign(email, process.env.JWT_SECRET);
             const newUser = UserModel({
                 name,
+                email,
                 address,
                 place,
                 country,
@@ -24,11 +27,9 @@ class AuthController {
                 lat,
                 province,
                 privilage: "user",
-                validToken: "",
+                validToken: token,
                 password: md5(password)
             });
-
-            //Generate token
 
             await newUser.save();
             return res.status(200).json({
@@ -42,7 +43,7 @@ class AuthController {
                 lat,
                 province
                 },
-                validToken
+                validToken: token
             });
         }
         catch(err) {
@@ -56,6 +57,7 @@ class AuthController {
         try {
             const user = await UserModel.findOne({email: email});
 
+            //TODO handle this through res
             if(!user) {
                 throw new Error("User not found.");
             }
@@ -64,11 +66,12 @@ class AuthController {
                 throw new Error("Invalid password.");
             }
             
-            //Generate token
+            const token = jwt.sign(email, process.env.JWT_SECRET);
+            UserModel.findOneAndUpdate({email: email}, {validToken: token});
 
             return res.status(200).json({
                 user: {...user, password: ""},
-                validToken: ""
+                validToken: token
             });
         }
         catch(err) {
@@ -81,13 +84,36 @@ class AuthController {
     }
 
     static async logout(req, res) {
-        //TODO
+        const {email, token} = req.body;
+
+        try {
+            const user = await UserModel.findOne({email: email});
+
+            if(!user) {
+                throw new Error("User not found.");
+            }
+
+            if(user.validToken != token) {
+                return res.status(401).json({
+                    errors: ["Access denied."]
+                });
+            }
+
+            await UserModel.findOneAndUpdate({email: email}, {validToken: ""});
+        }
+        catch(err) {
+            handleError(res, [err.message], "Logout");
+        }
     }
 
     static async deleteAccount(req, res) {
-        //Invalidate token
-
         try {
+            const userToken = (await UserModel.findOne({email: email})).validToken;
+            if(userToken != req.body.token) {
+                return res.status(401).json({
+                    errors: ["Access denied."]
+                });
+            }
             await UserModel.findOneAndDelete({email: req.body.email});
         }
         catch(err) {
